+++ packages/vueginate/src/components/VueginateBulma.vue
@@ -47,6 +47,10 @@ defineProps({
       return {}
     },
   },
+  fixedLength: {
+    type: Boolean,
+    default: true,
+  },
 })
 
 function changePage(page: number) {
@@ -74,6 +78,7 @@ function changePage(page: number) {
     }"
     :custom-styles="customStyles"
     :visible-always="visibleAlways"
+    :fixed-length="fixedLength"
     @page-change="changePage"
   >
     <template #previous="{ item }">