+++ packages/vueginate/src/components/VueginateTailwind.vue
@@ -45,6 +45,10 @@ defineProps({
       return {}
     },
   },
+  fixedLength: {
+    type: Boolean,
+    default: true,
+  },
 })
 
 function changePage(page: number) {
@@ -80,6 +84,7 @@ function changePage(page: number) {
     }"
     :custom-styles="customStyles"
     :visible-always="visibleAlways"
+    :fixed-length="fixedLength"
     @page-change="changePage"
   />
 </template>